Declare kFirstContainerResetPaint under correct #ifdef
All its uses happen under #if defined(USE_NEVA_APPRUNTIME),
but it was declared under #if defined(OS_WEBOSE).

Fixes compilation error when OS_WEBOSE is not defined and
USE_NEVA_APPRUNTIME is.

| ./../../git/src/third_party/blink/renderer/core/paint/first_meaningful_paint_detector.cc:267:19: error: 'kFirstContainerRes\
etPaint' is not a member of 'blink::PaintEvent'; did you mean 'kFirstContentfulPaint'?
|   267 |       PaintEvent::kFirstContainerResetPaint,
|       |                   ^~~~~~~~~~~~~~~~~~~~~~~~~
|       |                   kFirstContentfulPaint
| ./../../git/src/third_party/blink/renderer/core/paint/first_meaningful_paint_detector.cc:271:23: error: 'kFirstContainerRes\
etPaint' is not a member of 'blink::PaintEvent'; did you mean 'kFirstContentfulPaint'?
|   271 |           PaintEvent::kFirstContainerResetPaint));
|       |                       ^~~~~~~~~~~~~~~~~~~~~~~~~
|       |                       kFirstContentfulPaint